simple-agent-template/pyproject.toml
[project]
name = "single-agent-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.0",
    "langchain-core>=1.2.5",
    "langchain-openai>=1.1.6",
    "langchain-text-splitters>=1.1.0",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "pydantic-settings>=2.12.0",
]
